# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EarnPearls API",
    "description": "Version 1 API for authentication, surveys, wallet, withdrawals, and administration.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ep_session",
        "description": "Opaque, revocable HttpOnly session cookie."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Session-bound token required for authenticated state changes."
      },
      "providerSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Provider-Signature",
        "description": "Provider-adapter-specific webhook signature."
      }
    },
...
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "displayName",
                  "countryCode"
                ],
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 254,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 12,
                    "maxLength": 128,
                    "type": "string"
                  },
                  "displayName": {
                    "minLength": 1,
                    "maxLength": 100,
                    "type": "string"
                  },
                  "countryCode": {
                    "minLength": 2,
                    "maxLength": 2,
                    "pattern": "^[A-Z]{2}$",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/sessions": {
      "get": {
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "current",
                      "userAgent",
                      "createdAt",
                      "lastSeenAt",
                      "expiresAt"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "current": {
                        "type": "boolean"
                      },
                      "userAgent": {
                        "type": "string"
                      },
                      "createdAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "lastSeenAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "expiresAt": {
                        "format": "date-time",
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/sessions/{sessionId}": {
      "delete": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "sessionId",
            "required": true
          }
        ],
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 254,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 1,
                    "maxLength": 128,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "authenticated",
                    "user"
                  ],
                  "properties": {
                    "authenticated": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "user": {
                      "type": "object",
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "countryCode",
                        "accountStatus",
                        "emailVerified",
                        "createdAt"
                      ],
                      "properties": {
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "email": {
                          "format": "email",
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "countryCode": {
                          "minLength": 2,
                          "maxLength": 2,
                          "pattern": "^[A-Z]{2}$",
                          "type": "string"
                        },
                        "accountStatus": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "active"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "limited"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "suspended"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "disabled"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "archived"
                              ]
                            }
                          ]
                        },
                        "emailVerified": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "format": "date-time",
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-email": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token"
                ],
                "properties": {
                  "token": {
                    "minLength": 40,
                    "maxLength": 200,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 254,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/password-reset/confirm": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "password"
                ],
                "properties": {
                  "token": {
                    "minLength": 40,
                    "maxLength": 200,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 12,
                    "maxLength": 128,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/session": {
      "get": {
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "authenticated",
                    "user",
                    "capabilities"
                  ],
                  "properties": {
                    "authenticated": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "user": {
                      "type": "object",
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "countryCode",
                        "accountStatus",
                        "emailVerified",
                        "createdAt"
                      ],
                      "properties": {
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "email": {
                          "format": "email",
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "countryCode": {
                          "minLength": 2,
                          "maxLength": 2,
                          "pattern": "^[A-Z]{2}$",
                          "type": "string"
                        },
                        "accountStatus": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "active"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "limited"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "suspended"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "disabled"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "archived"
                              ]
                            }
                          ]
                        },
                        "emailVerified": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "format": "date-time",
                          "type": "string"
                        }
                      }
                    },
                    "capabilities": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout-all": {
      "post": {
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Current origin"
    }
  ],
...
}